--- src/Footer/index.scss
@@ -1,20 +1,0 @@
-footer {
-  width: 100%;
-  height:100px;
-  background-color: #919191;
-  text-align: center;
-}
-
-img {
-  width:15px;
-}
-
-.contact ul {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  list-style: none;
-}
-
-
-
