--- src/Header/index.scss
@@ -1,17 +1,0 @@
-header {
-  position: relative;  
-  width: 100%;
-  height: 100px;
-  background-color: #6a6a6a;
-  justify-content: center;
-}
-
-.logo {
-  position: absolute;
-  left: 5px;
-  top: 5px;
-  width: 90px;
-  height: 90px;
-  border-radius: 30%;;
-  background-color: rgb(250, 0, 0);
-}
