--- src/Navigation/VendorNavBar/VendorNavBar.scss
@@ -1,107 +1,0 @@
-:root {
-    --clr: #9ED5C5;
-}
-
-img {
-    width:45px;
-}
-
-.navigation {
-    position: relative;
-    width: 100%;
-    height: 50px;
-    background: #9ED5C5;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 10px;
-}
-
-.navigation ul {
-    display: flex;
-    width: 100%;
-}
-
-.navigation ul li {
-    position: relative;
-    list-style: none;
-    width: 150px;
-    height: 45px;
-    z-index: 1;
-}
-
-.navigation ul li a {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    width: 100%;
-    text-align: center;
-    font-weight: 500;
-}
-
-.navigation ul li a .icon {
-    position: relative;
-    display: block;
-    line-height: 50px;
-    font-size: 1.5em;
-    text-align: center;
-    transition: 0.5s;
-    color: var(--clr);
-}
-
-.navigation ul li.active a .icon {
-    transform: translateY(33px); 
-}
-
-.navigation ul li.active:hover a .icon {
-    transform: translateY(33px); 
-}
-
-.navigation ul li:hover a .icon {
-    transform: translateY(3px); 
-}
-
-.navigation ul li a .text {
-    position: absolute;
-    color: black;
-    font-weight:bold;
-    letter-spacing: 0.05em;
-    transition: .45s;
-    opacity: 0;
-    transform: translateY(-50px);
-}
-
-.navigation ul li.active a .text {
-    opacity: 1;
-    transform: translateY(-20px);
-}
-
-.indicator {
-    position: absolute;
-    top: 50%;
-    left: 79px;
-    width: 60px;
-    height: 60px;
-    background: #DEF5E5;
-    border-radius: 40%;
-    border: 6px solid var(--clr);
-    transition: .3s; 
-}
-
-.navigation ul li:nth-child(1).active ~ .indicator {
-    transform: translateX(calc(150px * 0));   
-}
-.navigation ul li:nth-child(2).active ~ .indicator {
-    transform: translateX(calc(150px * 1));
-}
-.navigation ul li:nth-child(3).active ~ .indicator {
-    transform: translateX(calc(150px * 2));
-}
-.navigation ul li:nth-child(4).active ~ .indicator {
-    transform: translateX(calc(150px * 3));
-}
-.navigation ul li:nth-child(5).active ~ .indicator {
-    transform: translateX(calc(150px * 4));
-}
